Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -8,6 +8,6 @@ char* welcome() {
 int main() {
     char* c = malloc(sizeof(char));
     scanf("%c", c);
-    printf("%d", *((int)c));
+    printf("%d",*c);
     return 0;
 }
